git-svn-id: http://192.168.20.11/sequoiadb/trunk@27923 6121f079-4a18-41d3-8c30-8870a322f2c2

+++ testcase_new/reliability/src/com/sequoiadb/netsplit/NetSplit2568.java
@@ -42,43 +42,41 @@ public class NetSplit2568 extends SdbTestBase {
     private GroupMgr groupMgr = null;
     private int totalCount;
     private String connectUrl;
+    private boolean clearFlag = false;
 
     @BeforeClass()
     public void setUp() {
-        Sequoiadb sdb = null;
+        Sequoiadb sdb = new Sequoiadb(coordUrl, "", "");
         try {
             System.out.println(
                     "the TestCase Name:" + this.getClass().getName() + ". the TestCase begin at:"
                             + new SimpleDateFormat("YYYY-MM-dd HH:mm:ss.SSS").format(new Date()));
             groupMgr = GroupMgr.getInstance();
-            
-            //CheckBusiness(true),检测当前集群环境，若存在异常返回false，
+
+            // CheckBusiness(true),检测当前集群环境，若存在异常返回false，
             if (!groupMgr.checkBusiness(true)) {
                 throw new SkipException("checkBusiness faile");
             }
-            
-            //确定切分的源和目标组
+
+            // 确定切分的源和目标组
             List<GroupWrapper> glist = groupMgr.getAllDataGroup();
             srcGroupName = glist.get(0).getGroupName();
             destGroupName = glist.get(1).getGroupName();
 
-            sdb = new Sequoiadb(coordUrl, "", "");
             CollectionSpace commCS = sdb.getCollectionSpace(csName);
             DBCollection cl = commCS.createCollection(clName,
                     (BSONObject) JSON
                             .parse("{ShardingKey:{'sk':1},Partition:4096,ShardingType:'hash',Group:'"
                                     + srcGroupName + "'}"));
-            //准备切分的数据
-            insertData(cl, 0, 1000);
+            // 准备切分的数据
+            insertData(cl, 0, 800);
         }
         catch (ReliabilityException e) {
             Assert.fail(this.getClass().getName() + " setUp error, error description:"
                     + e.getMessage() + "\r\n" + Utils.getStackString(e));
         }
         finally {
-            if (sdb != null) {
-                sdb.disconnect();
-            }
+            sdb.disconnect();
         }
     }
 
@@ -91,9 +89,7 @@ public class NetSplit2568 extends SdbTestBase {
             lob.write(id.getBytes());
             lob.close();
         }
-        synchronized (this) {
-            totalCount = totalCount + end - begin;
-        }
+        totalCount = totalCount + end - begin;
     }
 
     @Test
@@ -107,7 +103,7 @@ public class NetSplit2568 extends SdbTestBase {
             String srcSlaHost = srcGroup.getSlave().hostName();
             String destPriHost = destGroup.getMaster().hostName();
             String cataPriHost = cataGroup.getMaster().hostName();
-            //srcSlaHost为断网主机，若目标组主节点，catalog主节点在这台机上，changePrimary将尝试至多10次的切主操作
+            // srcSlaHost为断网主机，若目标组主节点，catalog主节点在这台机上，changePrimary将尝试至多10次的切主操作
             if (srcSlaHost.equals(destPriHost) && !destGroup.changePrimary(10)) {
                 throw new SkipException(destGroup.getGroupName() + " reelect faile");
             }
@@ -119,7 +115,7 @@ public class NetSplit2568 extends SdbTestBase {
             connectUrl = Utils.getDiffHostWithSvc(srcSlaHost, groupMgr.getAllHosts());
 
             // 建立并行任务
-            FaultMakeTask faultTask = BrokenNetwork.getFaultMakeTask(srcSlaHost, 0, 10, 10);
+            FaultMakeTask faultTask = BrokenNetwork.getFaultMakeTask(srcSlaHost, 0, 10, 15);
             TaskMgr mgr = new TaskMgr(faultTask);
             mgr.addTask(new Split("Split"));
             mgr.addTask(new Insert("insert"));
@@ -128,32 +124,43 @@ public class NetSplit2568 extends SdbTestBase {
             mgr.join();
             mgr.fini();
 
+            // TaskMgr检查线程异常
+            Assert.assertEquals(mgr.isAllSuccess(), true, mgr.getErrorMsg());
+
             // 最长等待20分钟的集群环境恢复
             Assert.assertEquals(Utils.checkBusinessWithTimeout(groupMgr, 1200), true,
                     "wait restore business faile");
 
-            // 结果校验
+            // 再次插入数据
             db = new Sequoiadb(connectUrl, "", "");
             db.setSessionAttr((BSONObject) JSON.parse("{PreferedInstance:'M'}"));
-            //TaskMgr检查线程异常
-            Assert.assertEquals(mgr.isAllSuccess(), true, mgr.getErrorMsg());
-            //源和目标数据量比对
+            DBCollection cl = db.getCollectionSpace(csName).getCollection(clName);
+            insertData(cl, 5000, 6000);
+
+            // 源和目标数据量比对
             long destCount = checkGroupData(db, destGroupName);
             long srcCount = checkGroupData(db, srcGroupName);
             Assert.assertEquals(srcCount + destCount, totalCount);
             destCount = checkGroupLob(db, destGroupName);
             srcCount = checkGroupLob(db, srcGroupName);
             Assert.assertEquals(srcCount + destCount, totalCount);
-            //组间一致性校验，尝试至多60次，每次间隔1秒
+            // 组间一致性校验，尝试至多60次，每次间隔1秒
             if (!destGroup.checkInspect(60, 1)) {
                 Assert.fail(destGroup.getInspectStdout());
             }
             if (!srcGroup.checkInspect(60, 1)) {
                 Assert.fail(srcGroup.getInspectStdout());
             }
+
+            clearFlag = true;
         }
         catch (ReliabilityException e) {
             Assert.fail(e.getMessage() + "\r\n" + Utils.getStackString(e));
+        }
+        finally {
+            if (db != null) {
+                db.disconnect();
+            }
         }
 
     }
@@ -225,16 +232,16 @@ public class NetSplit2568 extends SdbTestBase {
     public void tearDown() {
         Sequoiadb sdb = new Sequoiadb(SdbTestBase.coordUrl, "", "");
         try {
-            CollectionSpace commCS = sdb.getCollectionSpace(csName);
-            commCS.dropCollection(clName);
+            if (clearFlag) {
+                CollectionSpace commCS = sdb.getCollectionSpace(csName);
+                commCS.dropCollection(clName);
+            }
         }
         catch (BaseException e) {
             Assert.fail(e.getMessage() + "\r\n" + Utils.getStackString(e));
         }
         finally {
-            if (sdb != null) {
-                sdb.disconnect();
-            }
+            sdb.disconnect();
             System.out.println(
                     "the TestCase Name:" + this.getClass().getName() + ". the TestCase end at:"
                             + new SimpleDateFormat("YYYY-MM-dd HH:mm:ss.SSS").format(new Date()));
@@ -262,7 +269,7 @@ public class NetSplit2568 extends SdbTestBase {
                 throw new FaultException(mk.toString());
             }
             if (rt == OperateTask.faultStatus.RESTOREFAILURE) {
-               throw new FaultException(rt.toString());
+                throw new FaultException(rt.toString());
             }
         }
     }
@@ -282,7 +289,6 @@ public class NetSplit2568 extends SdbTestBase {
                 DBCollection cl = sdb.getCollectionSpace(csName).getCollection(clName);
                 cl.split(srcGroupName, destGroupName, (BSONObject) JSON.parse("{Partition:2048}"), // 切分
                         (BSONObject) JSON.parse("{Partition:4096}"));
-                insertData(cl, 5000, 8000);
             }
             catch (BaseException e) {
                 throw e;
@@ -296,7 +302,7 @@ public class NetSplit2568 extends SdbTestBase {
 
         @Override
         public void faultNotify(BSONObject status) {
-            
+
         }
 
     }
